requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
pydub==0.25.1
gTTS==2.4.0
smtplib-ssl==1.0.0
email-validator==2.1.0
flask==3.0.0
flask-session==0.5.0
werkzeug==3.0.1
jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7